Move the pytest marks to their own module
