+++ conftest.py
@@ -1,4 +1,4 @@
-import pytest
 def pytest_addoption(parser):
     parser.addoption("--runexternals", action="store_true",
         help="run tests that depend on external services")
+
