+++ python/Ganga/testlib/__init__.py
@@ -1,6 +1,1 @@
-import pytest
 
-external = pytest.mark.skipif(
-    not pytest.config.getoption("--runexternals"),
-    reason="need --runexternals option to run external tests"
-)
