+++ python/Ganga/testlib/mark.py
@@ -1,0 +1,6 @@
+import pytest
+
+external = pytest.mark.skipif(
+    not pytest.config.getoption("--runexternals"),
+    reason="need --runexternals option to run external tests"
+)
