+++ python/GangaLHCb/test/GPI/TestDataFiles.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from Ganga.testlib import external
+from Ganga.testlib.mark import external
 from Ganga.testlib.GangaUnitTest import GangaUnitTest
